Lot's of refactoring for thread/fork safety
Split out a separate DbConn object class to handle and track the
connection to the sqlite db file, with support for recovery, reconnect,
and auto new inc filename generation.

Additionally, the code is now written to automatically split off into
new db files for new/separate pids which happen with multiple
workers/threads. Since we already intended to merge log dbs later on, we
are now intentionally keeping each worker separate for maximum speed,
instead of opening and closing the connection with each write, which is
what would be needed probably, and might not even be threadsafe anyway,
since multiple requests can happen at once, and the last thing in the
world we want to do is inadvertantly block and turning performance back
to being essentially single-threaded

IN PROGRESS...